+++ grpc/lib/grpc/client/application.ex
@@ -3,6 +3,18 @@ defmodule GRPC.Client.Application do
   use Application
 
   def start(_type, _args) do
+    table = GRPC.Client.LoadBalancing
+
+    if :ets.whereis(table) == :undefined do
+      :ets.new(table, [
+        :named_table,
+        :public,
+        :set,
+        {:read_concurrency, true},
+        {:write_concurrency, true}
+      ])
+    end
+
     children = [
       {Registry, [keys: :unique, name: GRPC.Client.Registry]},
       {DynamicSupervisor, [name: GRPC.Client.Supervisor]}